+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+
+# 1.16.0
 * Upgrade `PyPDF2` 2.x to `pypdf` 5.0.1 (new name, same package)
 * Add `--image-quality` option to `combine_pdfs` tool